feat(fish/nixos): Add keychain to manage ssh agents
This commit introduces 'keychain' to the development environment to persist ssh-agent sessions across shells and reboots.

The 'keychain' package is added to the 'nixos/programs/cli.nix' file, and the 'fish/config.fish' is updated to start the keychain service.

This change simplifies the management of SSH keys and avoids having to repeatedly enter passphrases.