refactor(nixos): Parameterize username in NixOS configuration
This commit refactors the NixOS configuration to replace the hardcoded username "mimame" with a `username` variable.

The `username` is defined in `nixos/hosts/narnia/configuration.nix` and passed as a parameter to the various modules that need it.

This change makes the configuration more modular and easier to reuse for different users.